Correctable Migration and start realize events

+++ Assets/Main/Map/MapParamFuncs/Examplars/MU_FoodAppropriated.cs
@@ -4,7 +4,7 @@ using WorldMapStrategyKit;
 public struct MU_FoodAppropriated : IUpdater {
     private int _regionIndex;
 
-    private int _population;
+    private float _population;
     private float _flora;
     private float _fauna;
     private float _hunters;
